requirements.txt
grpcio
grpcio-tools
fastapi
uvicorn[standard]
python-dotenv
requests
